several improvements
