--- src/main/java/com/devteam45ldm/ldm/parser/Csv2Json.java
@@ -1,16 +1,0 @@
-package com.devteam45ldm.ldm.parser;
-
-import com.devteam45ldm.ldm.parser.templates.importDataStructures.ImportTemplate;
-
-public class Csv2Json {
-    private ImportTemplate importTemplate;
-
-    public Csv2Json(ImportTemplate importTemplate) {
-        this.importTemplate = importTemplate;
-    }
-
-    public String convert(String csv) {
-        // Implement CSV to JSON conversion logic
-        return importTemplate.toJson();
-    }
-}
